Structure: Extract styles for navigation
