+++ src/components/general/navigation/PescaNavigator/PescaNavigator/PescaNavigator.style.ts
@@ -1,0 +1,12 @@
+import { StyleSheet } from 'react-native';
+
+export const usePescaNavigatorStyles = () => {
+  return StyleSheet.create({
+    screensContainer: {
+      flexDirection: 'row',
+      width: '100%',
+      maxHeight: '100%',
+      paddingBottom: 35,
+    },
+  });
+};
