+++ src/components/general/navigation/PescaNavigator/PescaScreen/PescaScreen.style.ts
@@ -1,0 +1,25 @@
+import { StyleSheet } from 'react-native';
+
+export const usePescaScreenStyles = () => {
+  return StyleSheet.create({
+    screenWrapper: {
+      flexDirection: 'row',
+      width: '100%',
+    },
+    dummyWrapper: {
+      width: '100%',
+    },
+    displayFlex: {
+      display: 'flex',
+    },
+    displayNone: {
+      display: 'none',
+    },
+    positionRelative: {
+      position: 'relative',
+    },
+    positionAbsolute: {
+      position: 'absolute',
+    },
+  });
+};
