+++ src/components/general/navigation/PescaTabIcon/PescaTabIcon.component.tsx
@@ -1,11 +1,10 @@
-import variables from '@moneyboy/config/variables';
-import { useStyle } from '@moneyboy/hooks/useStyle';
 import { BottomTabNavigationEventMap } from '@react-navigation/bottom-tabs/lib/typescript/src/types';
 import { NavigationHelpers } from '@react-navigation/core';
 import { ParamListBase } from '@react-navigation/routers';
 import React from 'react';
-import { Insets, StyleSheet, TouchableWithoutFeedback, View } from 'react-native';
+import { Insets, TouchableWithoutFeedback, View } from 'react-native';
 import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons';
+import { usePescaTabIconStyles } from './PescaTabIcon.style';
 
 type PescaTabProps = {
   navigation: NavigationHelpers<ParamListBase, BottomTabNavigationEventMap>;
@@ -23,31 +22,7 @@ export const PescaTabIcon: React.FC<PescaTabProps> = ({ name, navigation, icon, 
     navigation.navigate(name);
   }
 
-  const { Tabs } = useStyle();
-  const styles = StyleSheet.create({
-    tab: {
-      minWidth: 32,
-      height: 32,
-      justifyContent: 'center',
-    },
-    icon: {
-      fontSize: variables.font.size.large,
-      color: Tabs.default.color,
-    },
-    focus: {
-      color: Tabs.focus.color,
-      shadowColor: Tabs.focus.shadow,
-      shadowRadius: 10,
-      shadowOffset: {
-        width: 1,
-        height: 1,
-      },
-      shadowOpacity: 0.6,
-    },
-    disabled: {
-      color: Tabs.disabled.color,
-    },
-  });
+  const styles = usePescaTabIconStyles();
 
   const hitStops: Insets = {
     top: 20,
